Remove HmdaPlatform from api project
